Migrating variables functions to the "state" namespace.
