+++ front/src/Api/iframe/room.ts
@@ -4,15 +4,11 @@ import { EnterLeaveEvent, isEnterLeaveEvent } from "../Events/EnterLeaveEvent";
 
 import {IframeApiContribution, queryWorkadventure, sendToWorkadventure} from "./IframeApiContribution";
 import { apiCallback } from "./registeredCallbacks";
-import {isSetVariableEvent, SetVariableEvent} from "../Events/SetVariableEvent";
 
 import type { ITiledMap } from "../../Phaser/Map/ITiledMap";
 
 const enterStreams: Map<string, Subject<EnterLeaveEvent>> = new Map<string, Subject<EnterLeaveEvent>>();
 const leaveStreams: Map<string, Subject<EnterLeaveEvent>> = new Map<string, Subject<EnterLeaveEvent>>();
-const setVariableResolvers = new Subject<SetVariableEvent>();
-const variables = new Map<string, unknown>();
-const variableSubscribers = new Map<string, Subject<unknown>>();
 
 interface TileDescriptor {
     x: number;
@@ -33,24 +29,6 @@ export const setMapURL = (url: string) => {
     mapURL = url;
 }
 
-export const initVariables = (_variables: Map<string, unknown>): void => {
-    for (const [name, value] of _variables.entries()) {
-        // In case the user already decided to put values in the variables (before onInit), let's make sure onInit does not override this.
-        if (!variables.has(name)) {
-            variables.set(name, value);
-        }
-    }
-
-}
-
-setVariableResolvers.subscribe((event) => {
-    variables.set(event.key, event.value);
-    const subject = variableSubscribers.get(event.key);
-    if (subject !== undefined) {
-        subject.next(event.value);
-    }
-});
-
 export class WorkadventureRoomCommands extends IframeApiContribution<WorkadventureRoomCommands> {
     callbacks = [
         apiCallback({
@@ -66,13 +44,6 @@ export class WorkadventureRoomCommands extends IframeApiContribution<Workadventu
             callback: (payloadData) => {
                 leaveStreams.get(payloadData.name)?.next();
             },
-        }),
-        apiCallback({
-            type: "setVariable",
-            typeChecker: isSetVariableEvent,
-            callback: (payloadData) => {
-                setVariableResolvers.next(payloadData);
-            }
         }),
     ];
 
@@ -118,31 +89,6 @@ export class WorkadventureRoomCommands extends IframeApiContribution<Workadventu
             data: tiles,
         });
     }
-
-    saveVariable(key : string, value : unknown): Promise<void> {
-        variables.set(key, value);
-        return queryWorkadventure({
-            type: 'setVariable',
-            data: {
-                key,
-                value
-            }
-        })
-    }
-
-    loadVariable(key: string): unknown {
-        return variables.get(key);
-    }
-
-    onVariableChange(key: string): Observable<unknown> {
-        let subject = variableSubscribers.get(key);
-        if (subject === undefined) {
-            subject = new Subject<unknown>();
-            variableSubscribers.set(key, subject);
-        }
-        return subject.asObservable();
-    }
-
 
     get id() : string {
         if (roomId === undefined) {
